Task should now ignore symbolic links with no destination

+++ task-fscopy/src/com/github/maybeec/oomph/task/fscopy/core/impl/FSCopyUtilImpl.java
@@ -133,6 +133,12 @@ public class FSCopyUtilImpl implements FSCopyUtil
 
       private FileVisitResult copy(Path fileOrDir) throws IOException
       {
+
+        if (Files.exists(fileOrDir))
+        {
+          return FileVisitResult.CONTINUE;
+        }
+
         try
         {
           Files.copy(fileOrDir, dstDir.resolve(srcDir.relativize(fileOrDir)));
